requirements.txt
python-telegram-bot[job-queue]==21.6
httpx==0.27.0
pytz==2024.1
